move class SubmittedVariantAccessioningService out of the eva package since this joins services from dbsnp and eva
